+++ app/src/androidTest/java/com/infinity/omos/ui/custom/OmosViewActions.kt
@@ -1,13 +1,11 @@
-package com.infinity.omos.ui.onboarding
+package com.infinity.omos.ui.custom
 
 import android.view.View
-import android.widget.EditText
 import androidx.test.espresso.UiController
 import androidx.test.espresso.ViewAction
 import androidx.test.espresso.matcher.ViewMatchers.isAssignableFrom
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import com.infinity.omos.ui.view.OmosFieldView
-import com.infinity.omos.utils.Pattern
 import org.hamcrest.Matcher
 import org.hamcrest.Matchers.allOf
